requirements.txt
eclipse-zenoh==1.7.2
fastapi==0.135.1
kuksa-client==0.5.1
python-dotenv==1.2.2
requests==2.32.5
uvicorn==0.41.0
